+++ test/tests_allocation.js
@@ -1,38 +1,55 @@
 // ------------------------------------------------------------
 QUnit.module('Assets allocation module', {
   before: function() {
-    // 
+    // Generate a random 2*2 covariance matrix 
+	this.randomCovarianceMatrix = function() {
+		// Generate a random 2*2 covariance matrix
+		var var1 = 1 - Math.random(); // belongs to ]0,1]
+		var var2 = 1 - Math.random(); // belongs to ]0,1]
+		var cov = 2*Math.random() - 1; // belongs to ]-1,1[
+		while (var1*var2 - cov*cov <= 0) { // The covariance matrix must be DP, so that the covariance cannot be any value: determinant must be strictly positive
+			cov = 2*Math.random() - 1;
+		}
+		
+		return [[var1, cov], [cov, var2]];
+	};
   }
 });
 
 
 QUnit.test('Equal weights portfolio', function(assert) {    
-  // Limit case, one asset
+  // Random data, n assets case
   {
-	  var weights = PortfolioAllocation.equalWeights(1);
-	  assert.equal(weights.length, 1, 'Equal weights with one asset');
-	  assert.equal(weights[0], 1, 'Equal weights with one asset #2');
-  }
-
-  // General case, generate random values for testing
-  {
-    var nbAssets = Math.floor(Math.random()*(50-2+1) + 2); // max 50 min 2
-    var weights = PortfolioAllocation.equalWeights(nbAssets);
+    // Generate a random number of assets
+	var nbAssets = Math.floor(Math.random()*(50-1+1) + 1); // max 50 min 1
+    
+	// Compute EW weights
+	var weights = PortfolioAllocation.equalWeights(nbAssets);
   
-    assert.equal(weights.length, nbAssets, 'Equal weights');
-    for (var j = 0; j < weights.length; ++j) {
-	  assert.equal(weights[j], 1/nbAssets, 'Equal weights #2');
+    // Check the number of output weights
+	assert.equal(weights.length, nbAssets, 'EW - Number of weights');
+	
+	// Compare EW weights to exact closed-form formula
+	var expectedWeights = [];
+	for (var i = 0; i < nbAssets; ++i) {
+		expectedWeights[i] = 1/nbAssets;
+	}
+    for (var i = 0; i < weights.length; ++i) {
+	  assert.equal(weights[i], expectedWeights[i], 'EW - Values ' + i);
 	}
   }  
 });
 
 
 QUnit.test('Equal risk budget portfolio', function(assert) {     
+  // Reference: Carvalho, Raul Leote de and Xiao, Lu and Moulin, Pierre, Demystifying Equity Risk-Based Strategies: A Simple Alpha Plus Beta Description
+  
   // Random data, n assets case
   {
-	var nbAssets = Math.floor(Math.random()*(50-3+1) + 3); // max 50 min 3
+	// Generate a random number of assets
+	var nbAssets = Math.floor(Math.random()*(50-1+1) + 1); // max 50 min 1
 	
-	// Generate n variances
+	// Generate n random variances
 	var sigma = new Array(nbAssets);
 	for (var i = 0; i < nbAssets; ++i) {
 		sigma[i] = 1 - Math.random(); // belongs to ]0,1];
@@ -41,13 +58,20 @@ QUnit.test('Equal risk budget portfolio', function(assert) {
 	// Compute ERB weights
 	var weights = PortfolioAllocation.equalRiskBudgetWeights(sigma);
 	
+	// Check the number of output weights
+	assert.equal(weights.length, nbAssets, 'ERB - Number of weights');
+	
 	// Compare ERB weights to exact closed-form formula, c.f. formula 3 of the reference
 	var denom = 0;
 	for (var i = 0; i < nbAssets; ++i) { 
 		denom += 1/Math.sqrt(sigma[i]); 
 	}
+	var expectedWeights = [];
 	for (var i = 0; i < nbAssets; ++i) {
-		assert.equal(Math.abs(weights[i] - 1/Math.sqrt(sigma[i]) / denom) <= 1e-8, true, 'ERB #1/' + i);  
+		expectedWeights[i] = 1/Math.sqrt(sigma[i]) / denom;
+	}
+	for (var i = 0; i < nbAssets; ++i) {
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'ERB - Values ' + i);  
 	}
   }
 });
@@ -61,7 +85,7 @@ QUnit.test('Equal risk contributions portfolio', function(assert) {
 	  var weights = PortfolioAllocation.equalRiskContributionWeights([[0.01, 0, 0, 0], [0, 0.04, 0, 0], [0, 0, 0.09, 0], [0, 0, 0, 0.16]]); 
 	  var expectedWeights =  [48, 24, 16, 12];
 	  for (var i = 0; i < expectedWeights.length; ++i) {
-		assert.equal(Math.round(100 * weights[i]), expectedWeights[i], 'ERC #1');
+		assert.equal(Math.round(100 * weights[i]), expectedWeights[i], 'ERC - Values #1 ' + i);
 	  }
   }
   
@@ -71,29 +95,27 @@ QUnit.test('Equal risk contributions portfolio', function(assert) {
 	  var weights = PortfolioAllocation.equalRiskContributionWeights([[0.01, 0.016, 0, 0], [0.016, 0.04, 0, 0], [0, 0, 0.09, -0.06], [0, 0, -0.06, 0.16]]); 
 	  var expectedWeights =  [38.4, 19.2, 24.3, 18.2];
 	  for (var i = 0; i < expectedWeights.length; ++i) {
-		  assert.equal(Math.round(1000 * weights[i])/10, expectedWeights[i], 'ERC #2');
+		  assert.equal(Math.round(1000 * weights[i])/10, expectedWeights[i], 'ERC - Values #2 ' + i);
 	  }
   }
   
   // Random data, 2 assets case, c.f. p 62 of reference
   {
-	  // Generate a 2*2 covariance matrix
-	  var var1 = 1 - Math.random(); // belongs to ]0,1]
-	  var var2 = 1 - Math.random(); // belongs to ]0,1]
-	  var cov = 2*Math.random() - 1; // belongs to ]-1,1[
-	  while (var1*var2 - cov*cov <= 0) { // The covariance matrix must be DP, so that the covariance cannot be any value: determinant must be strictly positive
-	    cov = 2*Math.random() - 1;
-	  }
+	  // Generate a random 2*2 covariance matrix
+	  var covMat = this.randomCovarianceMatrix();
 	  
 	  // Compute ERC weights
-	  var weights = PortfolioAllocation.equalRiskContributionWeights([[var1, cov], [cov, var2]]);
+	  var weights = PortfolioAllocation.equalRiskContributionWeights(covMat);
 	  
 	  // Compare ERC weights to exact closed-form formula, c.f. p 62 of reference
+	  var var1 = covMat[0][0];
+	  var var2 = covMat[1][1];
 	  var std1 = Math.sqrt(var1);
 	  var std2 = Math.sqrt(var2);
 	  var expectedWeights =  [(1/std1) / (1/std1 + 1/std2), (1/std2) / (1/std1 + 1/std2)];
-	  assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-8, true, 'ERC #3/1 ' + i);  
-	  assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-8, true, 'ERC #3/2 ' + i);
+	  for (var i = 0; i < expectedWeights.length; ++i) {
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'ERC - Values #3 ' + i);  
+	  }
   }
 });
 
@@ -103,38 +125,38 @@ QUnit.test('Risk budgeting portfolio', function(assert) {
     
     // Random data, 2 assets case, c.f. p 6 of reference
 	{
-		// Generate a 2*2 covariance matrix
-		var var1 = 1 - Math.random(); // belongs to ]0,1]
-		var var2 = 1 - Math.random(); // belongs to ]0,1]
-		var cov = 2*Math.random() - 1; // belongs to ]-1,1[
-		while (var1*var2 - cov*cov <= 0) { // The covariance matrix must be DP, so that the correlation cannot be any value: determinant must be strictly positive
-		  cov = 2*Math.random() - 1;
-		}
+	    // Generate a random 2*2 covariance matrix
+	    var covMat = this.randomCovarianceMatrix();
 	  
-		// Generate risk budgets
+		// Generate random risk budgets
 		var b = 1 - Math.random(); // belongs to ]0,1]
 
 		// Compute RB weights
-		var weights = PortfolioAllocation.riskBudgetingWeights([[var1, cov], [cov, var2]], [b, 1 - b]);
+		var weights = PortfolioAllocation.riskBudgetingWeights(covMat, [b, 1 - b]);
 		
 		// Compare RB weights to exact closed-form formula, c.f. p 6 of reference
+	    var var1 = covMat[0][0];
+	    var var2 = covMat[1][1];
+		var cov = covMat[0][1];
 		var std1 = Math.sqrt(var1);
 		var std2 = Math.sqrt(var2);
 		var corr = cov/(std1*std2);
 		var w_star = (b - 0.5)*corr*std1*std2 - b*var2 + std1*std2*Math.sqrt( (b - 0.5)*(b - 0.5)*corr*corr + b*(1 - b) );
 		w_star /= (1 - b)*var1 - b*var2 + 2*(b - 0.5)*corr*std1*std2;	
 		var expectedWeights = [w_star, 1 - w_star];
-		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-8, true, 'RB #1/1');  
-		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-8, true, 'RB #1/2');
+		for (var i = 0; i < expectedWeights.length; ++i) {
+			assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'RB - Values #1 ' + i);  
+		}
 	}
 	
 	
     // Random data, n assets case, 0 correlation, c.f. p 7 of reference
     // The associated matrix is always definite positive as long as all variances are strictly positive
 	{
-		var nbAssets = Math.floor(Math.random()*(20-3+1) + 3); // max 20 min 3
+		// Generate a random number of assets
+		var nbAssets = Math.floor(Math.random()*(20-2+1) + 2); // max 20 min 2
 		
-		// Generate a n*n covariance matrix with 0 correlation
+		// Generate a random n*n covariance matrix with 0 correlation
 		var sigma = new Array(nbAssets);
 		for (var i = 0; i < nbAssets; ++i) {
 			sigma[i] = new Array(nbAssets);
@@ -164,7 +186,7 @@ QUnit.test('Risk budgeting portfolio', function(assert) {
 			denom += Math.sqrt(rb[i]) * 1/Math.sqrt(sigma[i][i]); 
 		}
 		for (var i = 0; i < nbAssets; ++i) {
-			assert.equal(Math.abs(weights[i] - Math.sqrt(rb[i]) * 1/Math.sqrt(sigma[i][i]) / denom) <= 1e-8, true, 'RB #2/' + i);  
+			assert.equal(Math.abs(weights[i] - Math.sqrt(rb[i]) * 1/Math.sqrt(sigma[i][i]) / denom) <= 1e-8, true, 'RB - Values #2 ' + i);  
 		}
 	}
 });
@@ -184,8 +206,8 @@ QUnit.test('Most diversified portfolio', function(assert) {
 		
 		// Compare MDP weights to values provided p. 9 of the reference
 		var expectedWeights = [0.33, 0.67];
-		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP #1/1');
-		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP #1/2');
+		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP - Values #1 0');
+		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP - Values #1 1');
 	}
 	
 	// Duplicate 3 assets is skipped, as not a real example (asset A duplication is a polico, and thus would never be selected)
@@ -201,8 +223,8 @@ QUnit.test('Most diversified portfolio', function(assert) {
 		
 		// Compare MDP weights to values provided p. 11 of the reference
 		var expectedWeights = [0.67, 0.33];
-		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP #2/1');
-		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP #2/2');
+		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP - Values #2 0');
+		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP - Values #2 1');
 	}
 	
 	// Polico invariance 3 assets case, p. 10 of the reference
@@ -216,9 +238,9 @@ QUnit.test('Most diversified portfolio', function(assert) {
 		
 		// Compare MDP weights to values provided p. 11 of the reference
 		var expectedWeights = [0.33, 0.67, 0];
-		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP #3/1');
-		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP #3/2');
-		assert.equal(Math.abs(weights[2] - expectedWeights[2]) <= 1e-2, true, 'MDP #3/3');
+		assert.equal(Math.abs(weights[0] - expectedWeights[0]) <= 1e-2, true, 'MDP - Values #3 0');
+		assert.equal(Math.abs(weights[1] - expectedWeights[1]) <= 1e-2, true, 'MDP - Values #3 1');
+		assert.equal(Math.abs(weights[2] - expectedWeights[2]) <= 1e-2, true, 'MDP - Values #3 2');
 	}
 	
 	
@@ -235,9 +257,9 @@ QUnit.test('Most diversified portfolio', function(assert) {
 
 		// Compare MDP weights to values provided p. 11 of the reference
 		var expectedWeights = [25.7, 25.7, 48.6];
-		assert.equal(Math.round(weights[0]*1000)/10, expectedWeights[0], true, 'MDP #4/1');
-		assert.equal(Math.round(weights[1]*1000)/10, expectedWeights[1], true, 'MDP #4/2');
-		assert.equal(Math.round(weights[2]*1000)/10,  expectedWeights[2], true, 'MDP #4/3');
+		assert.equal(Math.round(weights[0]*1000)/10, expectedWeights[0], true, 'MDP - Values #4 0');
+		assert.equal(Math.round(weights[1]*1000)/10, expectedWeights[1], true, 'MDP - Values #4 1');
+		assert.equal(Math.round(weights[2]*1000)/10,  expectedWeights[2], true, 'MDP - Values #4 2');
 	}
 	
 	
@@ -259,14 +281,14 @@ QUnit.test('Most diversified portfolio', function(assert) {
 			// Compare MDP weights to values obtained through Matlab CVX library usage
 			var expectedWeights = [0.334077696834404, 0.000000171869935, 0.332780568181053, 0.333141372759862, 0.000000190356559];
 			for (var i = 0; i < nbAssets; ++i) {
-				assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-6, true, 'MDP #5/' + i);
+				assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-6, true, 'MDP - Values #5 ' + i);
 			}
 		}
 		
 		{
 			// In case of a covariance matrix, maximizing the diversification ratio
-			// is equivalent to minimizing the variance, using the correlation matrix,
-			// plus rescaling the obtained weights by the assets volatilities + normalizing
+			// is equivalent to minimizing the variance using the correlation matrix,
+			// plus rescaling the obtained weights by the assets volatilities and normalizing
 			// them if no cash, c.f. discussion above formula 4 of the reference.
 			
 			// Generate random standard deviations
@@ -300,7 +322,7 @@ QUnit.test('Most diversified portfolio', function(assert) {
 			
 			// Compare MDP weights to expected weights
 			for (var i = 0; i < nbAssets; ++i) {
-				assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-4, true, 'MDP #6/' + i);
+				assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-4, true, 'MDP - Values #6 ' + i);
 			}
 		}
 	}
@@ -325,29 +347,26 @@ QUnit.test('Minimum correlation heuristic portfolio', function(assert) {
 	  // Compare MinCorr weights to expected weights
 	  var expectedWeights = [0.21, 0.31, 0.48];
 	  for (var i = 0; i < nbAssets; ++i) { 
-		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-2, true, 'MinCorr weights ' + i);
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-2, true, 'MinCorr - Values #1 ' + i);
 	  }
   }
 
   // Limit case: 2 assets, must be equal to ERB
   {
-		// Generate a 2*2 covariance matrix
-		var var1 = 1 - Math.random(); // belongs to ]0,1]
-		var var2 = 1 - Math.random(); // belongs to ]0,1]
-		var cov = 2*Math.random() - 1; // belongs to ]-1,1[
-		while (var1*var2 - cov*cov < 0) { // The covariance matrix must be DSP, so that the correlation cannot be any value: determinant must be positive
-		  cov = 2*Math.random() - 1;
-		}
+	    // Generate a random 2*2 covariance matrix
+	    var covMat = this.randomCovarianceMatrix();
 		
 		// Compute MinCorr weights
-		var weights = PortfolioAllocation.minCorrWeights([[var1, cov], [cov, var2]]);
+		var weights = PortfolioAllocation.minCorrWeights(covMat);
 		
 		// Compute ERB weights
+	    var var1 = covMat[0][0];
+	    var var2 = covMat[1][1];
 		var expectedWeights = PortfolioAllocation.equalRiskBudgetWeights([var1, var2]);
 		
 		// Compare the weights
 		for (var i = 0; i < 2; ++i) { 
-			assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'MinCorr weights 2 assets ' + i);
+			assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'MinCorr - 2 assets ' + i);
 		}
   }  
 });
@@ -368,7 +387,141 @@ QUnit.test('Proportional minimum variance heuristic portfolio', function(assert)
 	  // Compare MVA weights to expected weights
 	  var expectedWeights = [0.18, 0.07, 0.07, 0.68];
 	  for (var i = 0; i < nbAssets; ++i) { 
-		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-2, true, 'MinVar weights ' + i);
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-2, true, 'MinVar - Values #1 ' + i);
 	  }
   }
 });
+
+
+
+QUnit.test('Cluster risk parity portfolio', function(assert) {    
+  // Error cases
+  {
+	// Dummy covariance matrix
+	var sigma = [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]];
+
+	// Unsupported clustering method
+    assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'none' }) },
+		new Error('unsupported clustering method'),
+		"CRP - Manual clustering checks unsupported clustering method");
+
+    // In case manual clusters are provided for n assets, they must form a partition of the set [1..n]
+	// Test no clusters
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual' }) }, 
+		new Error('missing asset index: 1'),
+		"CRP - Manual clustering checks no clusters");
+	
+	// Test the presence of an empty cluster
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual', clusters: [[1, 2, 3, 4], []] }) }, 
+		new Error('empty cluster at index: 1'),
+		"CRP - Manual clustering checks empty cluster");
+	
+	// Test asset index out of bounds #1
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual', clusters: [[0, 1, 2, 3]] }) }, 
+		new Error('asset index out of bounds: 0'),
+		"CRP - Manual clustering checks out of bounds #1");
+		
+	// Test asset index out of bounds #2
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual', clusters: [[1, 2, 3, 5]] }) }, 
+		new Error('asset index out of bounds: 5'),
+		"CRP - Manual clustering checks out of bounds #2")
+
+	// Test missing asset index
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual', clusters: [[1, 2, 3]] }) }, 
+		new Error('missing asset index: 4'),
+		"CRP - Manual clustering checks missing asset index")
+	
+	// Test duplicate asset index
+	assert.throws(function() { 
+		PortfolioAllocation.clusterRiskParityWeights(sigma, { clusteringMode: 'manual', clusters: [[1, 1, 2, 3]] }) }, 
+		new Error('duplicate asset index: 1'),
+		"CRP - Manual clustering checks duplicate asset index")
+  } 
+
+  
+  // Limit cases
+  {
+	// Example taken from the ERC
+	var sigma = [[0.01, 0.016, 0, 0], [0.016, 0.04, 0, 0], [0, 0, 0.09, -0.06], [0, 0, -0.06, 0.16]];
+	var ercWeights = [38.4, 19.2, 24.3, 18.2];
+
+	// Limit case #1: in case all assets belong to a unique cluster, the CRP weights must be equal to the ERC weights
+	{
+		var opt = { clusteringMode: 'manual', clusters: [[1, 2, 3, 4]] };
+		var weights = PortfolioAllocation.clusterRiskParityWeights(sigma, opt); 
+		var expectedWeights = ercWeights;
+		for (var i = 0; i < expectedWeights.length; ++i) {
+			assert.equal(Math.round(1000 * weights[i])/10, expectedWeights[i], 'CRP - Values #1 ' + i);
+		}
+	}
+  
+	  // Limit case #2: in case each assets belong to one cluster, the CRP weights must be equal to the ERC weights
+	  {
+		  var opt = { clusteringMode: 'manual', clusters: [[1], [2], [3], [4]] };
+		  var weights = PortfolioAllocation.clusterRiskParityWeights(sigma, opt); 
+		  var expectedWeights = ercWeights;
+		  for (var i = 0; i < expectedWeights.length; ++i) {
+			  assert.equal(Math.round(1000 * weights[i])/10, expectedWeights[i], 'CRP - Values #2 ' + i);
+		  }
+	  }
+  }
+  
+  
+  // In case of 2 clusters of 2 assets each, the final weights must verify w_1f = p * w_1erc, w_2f = p * w_2erc,
+  // w_3f = (1-p) * w_3erc, w_4f = (1-p) * w_4erc with w_ierc the weight for asset i obtained though the ERC applied
+  // intra clusters and p a real number belonging to ]0,1[ dependant on the clusters variances/correlation.
+  {
+      // Generate two random 2*2 covariance matrices, one for each of the two clusters
+	  var sigma_c1 = this.randomCovarianceMatrix();
+	  var sigma_c2 = this.randomCovarianceMatrix();
+
+	  // Create a 4x4 bloc matrix using the two 2x2 matrices above
+	  var sigma = [[sigma_c1[0][0], sigma_c1[0][1], 0, 0], [sigma_c1[1][0], sigma_c1[1][1], 0, 0], [0, 0, sigma_c2[0][0], sigma_c2[0][1]], [0, 0, sigma_c2[1][0], sigma_c2[1][1]]];
+		
+      // Compute CRP weights for the whole portfolio
+	  var opt = { clusteringMode: 'manual', clusters: [[1,2], [3,4]] };
+	  var weights = PortfolioAllocation.clusterRiskParityWeights(sigma, opt); 
+	  
+  	  // Compute ERC weights for each cluster
+	  var weights_c1 = PortfolioAllocation.equalRiskContributionWeights(sigma_c1); 
+	  var weights_c2 = PortfolioAllocation.equalRiskContributionWeights(sigma_c2); 
+
+	  // Ensure the relationships described above are satisfied
+	  var p = weights[0]/weights_c1[0];
+	  assert.equal(Math.abs(p - weights[1]/weights_c1[1]) <= 1e-8, true, 'CRP - Values #3 0');
+	  assert.equal(Math.abs(1-p - weights[2]/weights_c2[0]) <= 1e-8, true, 'CRP - Values #3 1');
+	  assert.equal(Math.abs(1-p - weights[3]/weights_c2[1]) <= 1e-8, true, 'CRP - Values #3 2');
+  }
+  
+
+  
+  // CRP using FTCA, using static data
+  {
+	// Example taken from the ERC
+	var sigma = [[0.01, 0.016, 0, 0], [0.016, 0.04, 0, 0], [0, 0, 0.09, -0.06], [0, 0, -0.06, 0.16]];
+	
+	// Call CRP with no options, which should defaults to FTCA
+	var expectedWeights = [ 0.3262379197337176, 0.16311896054014643, 0.29179607083498427, 0.21884704889115175 ];
+	var weights = PortfolioAllocation.clusterRiskParityWeights(sigma);
+	for (var i = 0; i < sigma[0].length; ++i) { 
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'CRP - Values #4 ' + i);
+	}
+	  
+	// Call CRP with FTCA mode, which should output the same result as above
+	var opt = { clusteringMode: 'ftca' };
+	weights = PortfolioAllocation.clusterRiskParityWeights(sigma, opt); 		
+	for (var i = 0; i < sigma[0].length; ++i) { 
+		assert.equal(Math.abs(weights[i] - expectedWeights[i]) <= 1e-8, true, 'CRP - Values #5 ' + i);
+	}
+
+  }
+});
+
+  
+